% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_legend.R
\name{align_legend}
\alias{align_legend}
\title{Align the legend keys relative to the legend title}
\usage{
align_legend(p, hjust = 0.5)
}
\arguments{
\item{p}{The plot on which the legend should be aligned}

\item{hjust}{Horizontal alignment parameter. 0 means left-aligned, 0.5 centered, and 1 right-aligned.}
}
\value{
A \code{gtable} object.
}
\description{
Align the legend keys relative to the legend title
}
\examples{
# single legend
p <- ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Petal.Width)) + geom_point()
ggdraw(align_legend(p)) # centered
ggdraw(align_legend(p), hjust = 1) # right aligned

# multiple legends
p2 <- ggplot(mtcars, aes(disp, mpg, fill = hp, shape = factor(cyl), size = wt)) + 
  geom_point(color = "white") +
  scale_shape_manual(values = c(23, 24, 21), name = "cylinders") +
  scale_fill_continuous_carto(palette = "Emrld", name = "power (hp)", breaks = c(100, 200, 300)) +
  xlab("displacement (cu. in.)") +
  ylab("fuel efficiency (mpg)") +
  guides(shape = guide_legend(override.aes = list(size = 4, fill = "#329D84")),
         size = guide_legend(override.aes = list(shape = 21, fill = "#329D84"),
                             title = "weight (1000 lbs)")) +
  theme_half_open() + background_grid()

# works but maybe not the expected result
ggdraw(align_legend(p2))

# more sensible layout
ggdraw(align_legend(p2 + theme(legend.position = "top", legend.direction = "vertical")))
}
