% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blue_jays.R
\docType{data}
\name{blue_jays}
\alias{blue_jays}
\title{Body measurements for a sample of blue jays

This dataset corresponds to the \code{BlueJays} dataset from the Stat2Data package.}
\format{An object of class \code{data.frame} with 123 rows and 9 columns.}
\usage{
blue_jays
}
\description{
Body measurements for a sample of blue jays

This dataset corresponds to the \code{BlueJays} dataset from the Stat2Data package.
}
\keyword{datasets}
