% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridges.R
\docType{data}
\name{bridges}
\alias{bridges}
\title{Pittsburgh bridges}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 108 rows and 13 columns.}
\usage{
bridges
}
\description{
Data on Pittsburgh bridges, collected by Yoram Reich and Steven J. Fenves, Carnegie Mellon University.
}
\examples{
library(tidyverse)

select(bridges, MATERIAL, ERECTED) \%>\%
  table() \%>\% 
  reshape2::melt() \%>\% 
  ggplot(aes(ERECTED, value, fill = MATERIAL)) + geom_col(position = "stack")
}
\references{
Provided by the UCI Machine Learning Repository at: \url{https://archive.ics.uci.edu/ml/datasets/Pittsburgh+Bridges}
}
\keyword{datasets}
