% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_examples.R
\docType{data}
\name{correlation_examples}
\alias{correlation_examples}
\title{Random variates with specific correlation coefficients}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 300 rows and 3 columns.}
\usage{
correlation_examples
}
\description{
This dataset contains six sets of 50 x-y pairs with correlation coefficients
0.2, 0.6, 0.9, -0.2, -0.6, -0.9. Each x and y variable is drawn from a standard
normal distribution.
}
\examples{
ggplot(correlation_examples, aes(x, y)) + 
  geom_point() + facet_wrap(~r)
}
\keyword{datasets}
