% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cows.R
\docType{data}
\name{cows}
\alias{cows}
\title{Butterfat percentage for cows

This dataset corresponds to the \code{Cows} dataset from the PASWR package. It was
taken from Sokal and Rohlf.}
\format{An object of class \code{data.frame} with 100 rows and 3 columns.}
\source{
The Canadian Record of Performance for Purebred Dairy Cattle
}
\usage{
cows
}
\description{
Butterfat percentage for cows

This dataset corresponds to the \code{Cows} dataset from the PASWR package. It was
taken from Sokal and Rohlf.
}
\keyword{datasets}
