% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dow_jones_industrial.R
\docType{data}
\name{dow_jones_industrial}
\alias{dow_jones_industrial}
\title{Dow Jones industrial average}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8515 rows and 6 columns.}
\source{
\url{finance.yahoo.com}
}
\usage{
dow_jones_industrial
}
\description{
This dataset contains daily opening, high, low, and closing prices of the
Dow Jones Industrial Average from 1985 to 2018. Data was downloaded
from Yahoo finance on 11/7/2018.
}
\examples{
ggplot(dow_jones_industrial, aes(date, close)) + geom_line()
}
\keyword{datasets}
