% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_land.R
\name{draw_land}
\alias{draw_land}
\alias{draw_ocean}
\title{Draw land masses in orthographic projection}
\usage{
draw_land(coord, cenlat = 40, cenlong = 15, xlim = c(-1.2, 1.2),
  ylim = c(-1.2, 1.2), col = "#E69F00B0", bg_col = "white",
  line_col = "black", lwd = 0.5)

draw_ocean(cenlat = 40, cenlong = 15, xlim = c(-1.2, 1.2),
  ylim = c(-1.2, 1.2), col = "#56B4E950", line_col = "grey30",
  lwd = 0.5)
}
\arguments{
\item{coord}{Dataframe holding polygon coordinates in columns x and y, representing
longitude and latitude. Polygons are separated by rows of NA.}

\item{cenlat}{Center latitude of projected earth.}

\item{cenlong}{Center longitude of projected earth.}

\item{xlim, ylim}{Limits of the enclosing plot. The actual earth is projected
from -1 to 1 in x and y.}

\item{col}{Land or ocean color.}

\item{bg_col}{Background color.}

\item{line_col}{Line color.}

\item{lwd}{Line width.}
}
\description{
The function \code{draw_land()} draws a set of polygons representing land masses.
The function \code{draw_ocean()} draws the ocean background and graticules.
}
\examples{
# static plot
world <- map_polys$world
cenlat <- 45
cenlong <- 0
draw_ocean(cenlat, cenlong)
draw_land(world, cenlat, cenlong)

\dontrun{
# animation
makeplot <- function() {
  cenlat <- 35
  by <- 2
  longs <- seq(-180, 180-by, by = by)
  for (cenlong in longs) {
    draw_ocean(cenlat, cenlong)
    draw_land(map_polys$usa, cenlat, cenlong, col = "#FF0000B0") 
    draw_land(map_polys$world_no_usa, cenlat, cenlong, col = "#C0C0C0B0")
  }
}
gifski::save_gif(makeplot(), "world.gif", delay = 1/24, 500, 500)
}
}
