% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forensic_glass.R
\docType{data}
\name{forensic_glass}
\alias{forensic_glass}
\title{Measurements of forensic glass fragments

This dataset corresponds to the \code{fgl} dataset from the MASS package.}
\format{An object of class \code{data.frame} with 214 rows and 10 columns.}
\source{
B. German, Central Research Establishment, Home Office Forensic Science Service, UK
}
\usage{
forensic_glass
}
\description{
Measurements of forensic glass fragments

This dataset corresponds to the \code{fgl} dataset from the MASS package.
}
\keyword{datasets}
