% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy.R
\docType{data}
\name{happy}
\alias{happy}
\title{Data related to happiness from the general social survey}
\format{An object of class \code{data.frame} with 51020 rows and 10 columns.}
\usage{
happy
}
\description{
The data is a small sample of variables related to happiness from the general social survey (GSS).
This dataset is equivalent to \code{happy} from the \code{productplots} package.
}
\keyword{datasets}
