% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_census.R
\docType{data}
\name{hawaii_income}
\alias{hawaii_income}
\title{Median income in Hawaii counties}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 5 columns.}
\usage{
hawaii_income
}
\description{
Median income in Hawaii counties for 2010 and 2015, from the 2010 and 2015 five-year American Community Survey.
}
\examples{
hawaii_income_diff <- select(hawaii_income, -median_income_moe) \%>\%
  spread(year, median_income) \%>\%
  mutate(income_diff = `2015` - `2010`)

ggplot(hawaii_income_diff, aes(county, income_diff)) + geom_col()
}
\keyword{datasets}
