% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_prices.R
\docType{data}
\name{house_prices}
\alias{house_prices}
\title{House Price Index and unemployment, by state}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25704 rows and 5 columns.}
\source{
House Price Index: \url{http://www.freddiemac.com/research/docs/states.xls}

Unemployment: \url{https://download.bls.gov/pub/time.series/la/la.txt}
}
\usage{
house_prices
}
\description{
This data set combines the Freddie Mac House Price Index with
unemployment data.
}
\details{
Data columns:
\itemize{
\item \code{date} Date (year and month only)
\item \code{state} The state
\item \code{unemploy_perc} Unemployment, in percent
\item \code{house_price_index} House Price Index, not seasonally adjusted
\item \code{house_price_perc} Year-over-year percent change in House Price Index.
}
}
\examples{
ggplot(filter(house_prices, state == "California", year(date) > 2000),
       aes(unemploy_perc, house_price_perc)) + geom_path()
}
\keyword{datasets}
