% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_by_age.R
\docType{data}
\name{income_by_age}
\alias{income_by_age}
\title{U.S. median income by age}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 42 rows and 4 columns.}
\usage{
income_by_age
}
\description{
2016 median income by age, and reported for all races and broken down by race. In general,
the non-exclusive data value (alone or in combination, "a.o.i.c.") was chosen.
}
\examples{


library(tidyverse)

income_by_age \%>\% filter(race == "all") \%>\%
  ggplot(aes(x = age, y = median_income)) +
    geom_col()

income_by_age \%>\% filter(race == "all") \%>\%
  ggplot(aes(x = fct_reorder(age, desc(median_income)), y = median_income)) +
    geom_col()

income_by_age \%>\% filter(race \%in\% c("white", "asian", "black", "hispanic")) \%>\%
  mutate(race = fct_relevel(race, c("asian", "white", "hispanic", "black"))) -> income_df
ggplot(income_df, aes(x = age, y = median_income, fill = race)) +
  geom_col(position = "dodge") +
  scale_fill_brewer(name = NULL) + theme_minimal_hgrid()

ggplot(income_df, aes(x = race, y = median_income, fill = age)) +
  geom_col(position = "dodge") +
  scale_fill_brewer(name = NULL) + theme_minimal_hgrid()

ggplot(income_df, aes(x = age, y = median_income)) +
  geom_col(position = "dodge") +
  facet_wrap(~race) + theme_minimal_hgrid()
}
\references{
United States Census Bureau

Downloaded on 02/11/2018 from: https://www.census.gov/data/tables/time-series/demo/income-poverty/historical-income-households.html
}
\keyword{datasets}
