% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legosets.R
\docType{data}
\name{legosets}
\alias{legosets}
\title{Information about every LEGO set manufactured from 1970 to 2015}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6172 rows and 14 columns.}
\source{
\url{https://github.com/seankross/lego}
}
\usage{
legosets
}
\description{
This dataset corresponds to the \code{legosets} dataset from Sean Kross's \code{lego} package.
The data were originally obtained from \url{https://brickset.com/}.
}
\keyword{datasets}
