% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_polys.R
\docType{data}
\name{map_polys}
\alias{map_polys}
\title{Various map polygons}
\format{An object of class \code{list} of length 3.}
\usage{
map_polys
}
\description{
All map polygons were obtained from the maps package (v3.3.0) via the
function \code{maps:::map.poly()}.
}
\examples{
cenlat <- 35
cenlong <- -135

draw_ocean(cenlat, cenlong)
draw_land(map_polys$usa, cenlat, cenlong, col = "#FF0000B0") 
draw_land(map_polys$world_no_usa, cenlat, cenlong, col = "#C0C0C0B0")
}
\seealso{
\code{\link[=draw_land]{draw_land()}}
}
\keyword{datasets}
