% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_land.R
\name{orthproj}
\alias{orthproj}
\title{Project lat/long to x-y in orthographic projection.}
\usage{
orthproj(lat, long, cenlat, cenlong, remove_back = FALSE)
}
\arguments{
\item{lat}{Vector of latitudes, in degrees.}

\item{long}{Vector of longitudes, in degrees}

\item{cenlat}{Center latitude of projected earth.}

\item{cenlong}{Center longitude of projected earth.}

\item{remove_back}{Boolean indicating whether the back of the globe
should be removed.}
}
\description{
This function projects latitude/longitude to an x--y plane (with ranges -1 to 1) using
an orthographic projection. It also determines whether points are in the front or at the
back of the globe, and it can remove the back.
}
