% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_plot.R
\name{palette_plot}
\alias{palette_plot}
\title{Plot a palette of colors, with or without color labels}
\usage{
palette_plot(colors, label_size = 6, print_color_labels = TRUE)
}
\arguments{
\item{colors}{Vector of colors}

\item{label_size}{Size of the labels}

\item{print_color_labels}{Should labels be printed or not?}
}
\description{
Plot a palette of colors, with or without color labels
}
\examples{
colors <- c("#E495A5", "#C7A76C", "#86B875", "#39BEB1", "#7DB0DD", "#CD99D8")
palette_plot(colors)
}
