% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_turbines.R
\docType{data}
\name{sfbay_maps}
\alias{sfbay_maps}
\title{Maps of the bay area in California}
\format{An object of class \code{list} of length 5.}
\usage{
sfbay_maps
}
\description{
Map tiles obtained from maps.stamen.com and from openstreetmap.org.
}
\examples{
library(ggmap)

ggmap(sfbay_maps$sfbay_bg) + inset_ggmap(sfbay_maps$sfbay_lines)
ggmap(sfbay_maps$shiloh_osm)
}
\keyword{datasets}
