% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp}
\alias{stamp}
\alias{stamp_good}
\alias{stamp_phantom}
\alias{stamp_bad}
\alias{stamp_sketchy}
\alias{stamp_ugly}
\title{Stamp plots with a label, such as good, bad, or ugly.}
\usage{
stamp(p, color, alpha, label, family = dviz_font_family_bold,
  fontface = "plain", clip = "on")

stamp_good(p, clip = "on")

stamp_phantom(p, clip = "on")

stamp_bad(p, clip = "on")

stamp_sketchy(p, clip = "on")

stamp_ugly(p, clip = "on")
}
\arguments{
\item{p}{The plot to stamp}

\item{color}{The color of the stamp}

\item{alpha}{Transparency level of the stamp}

\item{label}{The text label used for the stamp}

\item{family}{Font family}

\item{fontface}{Font face}

\item{clip}{Should figure be clipped (default is "on")}
}
\description{
Stamp plots with a label, such as good, bad, or ugly.
}
\examples{
library(ggplot2)

p <- ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width, 
                      color = factor(Petal.Width))) +
  geom_point()
  
stamp_bad(p + guides(color = "none"))
stamp_ugly(p)
}
