% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tech_stocks.R
\docType{data}
\name{tech_stocks}
\alias{tech_stocks}
\title{Stock prices of four major tech companies}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9569 rows and 6 columns.}
\usage{
tech_stocks
}
\description{
This dataset contains several years of daily closing prices for four major
tech companies, Apple, Google, Microsoft, and Facebook.
}
\examples{
ggplot(tech_stocks, aes(x=date, y=price_indexed, color=ticker)) +
  geom_line() +
  scale_color_manual(values=c("#000000", "#E69F00", "#56B4E9", "#009E73"),
                     name="",
                     breaks=c("FB", "GOOG", "MSFT", "AAPL"),
                     labels=c("Facebook", "Alphabet", "Microsoft", "Apple")) +
  scale_x_date(name="year",
               limits=c(ymd("2012-06-01"), ymd("2017-05-31")),
               expand=c(0,0)) + 
  scale_y_continuous(name="stock price, indexed",
                     limits = c(0, 560),
                     expand=c(0,0)) +
  theme_minimal_hgrid()
}
\keyword{datasets}
