% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_census.R
\docType{data}
\name{texas_race}
\alias{texas_race}
\title{Racial composition in Texas counties}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1016 rows and 7 columns.}
\usage{
texas_race
}
\description{
Percent (\code{pct}) of White, Black, Asian, and Hispanic inhabitants in Texas counties.
Data were obtained from the 2010 decennial US Census.
}
\examples{
library(sf)

texas_race \%>\%
filter(variable == "White") \%>\%
  ggplot(aes(fill = pct)) +
  geom_sf(color = "white") +
  coord_sf(datum = NA) + 
  theme_map() +
  #scale_fill_continuous_diverging(palette = "Blue-Red 3", # uncomment for alternative color scale
  scale_fill_continuous_carto(palette = "Earth",
                              mid = 50,
                              limits = c(0, 100),
                              breaks = 25*(0:4),
                              name = "\% White") +
  theme(legend.title.align = 0.5,
        legend.text.align = 0)
}
\keyword{datasets}
