% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_half_open}
\alias{theme_half_open}
\title{Half-open theme}
\usage{
theme_half_open(font_size = 14, font_family = "", line_size = 0.5)
}
\arguments{
\item{font_size}{Theme font size}

\item{font_family}{Theme font family}

\item{line_size}{Theme line size}
}
\description{
Plain theme with left and bottom axis lines and no grid lines.
}
