% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_minimal_grid}
\alias{theme_minimal_grid}
\title{Theme with horizontal and vertical grid lines}
\usage{
theme_minimal_grid(font_size = 14, font_family = "", line_size = 0.5)
}
\arguments{
\item{font_size}{Theme font size}

\item{font_family}{Theme font family}

\item{line_size}{Theme line size}
}
\description{
Minimal theme with horizontal and vertical grid lines.
}
