% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_minimal_hgrid}
\alias{theme_minimal_hgrid}
\title{Theme with horizontal grid lines}
\usage{
theme_minimal_hgrid(font_size = 14, font_family = "", line_size = 0.5)
}
\arguments{
\item{font_size}{Theme font size}

\item{font_family}{Theme font family}

\item{line_size}{Theme line size}
}
\description{
Minimal theme with horizontal grid lines. Useful for plots where the main
comparison of interest happens along the y axis.
}
