% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_census.R
\docType{data}
\name{US_census}
\alias{US_census}
\title{US census data at the county level}
\format{An object of class \code{data.frame} with 3143 rows and 53 columns.}
\usage{
US_census
}
\description{
Data for 3143 counties in the United States. This dataset is equivalent to \code{countyComplete} from the \code{openintro} package.
}
\examples{
library(dplyr)
library(ggplot2)

data <- left_join(US_census, US_regions) \%>\%
 group_by(region, division, state) \%>\%
 summarize(pop2000 = sum(pop2000, na.rm = TRUE),
           pop2010 = sum(pop2010, na.rm = TRUE),
           popgrowth = (pop2010-pop2000)/pop2000,
           area = sum(area)) \%>\%
 arrange(popgrowth) \%>\%
 ungroup() \%>\%
 mutate(state = factor(state, levels = state),
        region = factor(region, levels = c("West", "South", "Midwest", "Northeast")))


ggplot(data, aes(x = state, y = popgrowth, fill = region)) + geom_col() + coord_flip()
}
\seealso{
\link{US_regions}
}
\keyword{datasets}
