% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_income_age.R
\docType{data}
\name{US_income_age}
\alias{US_income_age}
\title{United States median income and age per county}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3220 rows and 6 columns.}
\usage{
US_income_age
}
\description{
The data frame \code{us_income_age} provides median income with margin of error (moe)
and median age with moe for each county in the U.S. The moe is a standard 90\%
confidence interval, as provided by the US Census.
}
\details{
Data was obtained from the US Census, 2015 ACS.
}
\examples{
ggplot(US_income_age, aes(age, income)) + 
  geom_errorbarh(aes(xmin = age - age.moe, xmax = age + age.moe), height = 0) +
  geom_errorbar(aes(ymin = income - income.moe, ymax = income + income.moe), width = 0) +
  geom_point(size = 0.2) +
  facet_wrap(~state)
  
filter(US_income_age, state == "Virginia") \%>\%
  ggplot(aes(age, income)) + 
  geom_errorbarh(aes(xmin = age - age.moe, xmax = age + age.moe), height = 0) +
  geom_errorbar(aes(ymin = income - income.moe, ymax = income + income.moe), width = 0) +
  geom_point()
}
\keyword{datasets}
