% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/US_regions.R
\docType{data}
\name{US_regions}
\alias{US_regions}
\title{US regions and divisions}
\format{An object of class \code{data.frame} with 51 rows and 4 columns.}
\usage{
US_regions
}
\description{
The definitions of US regions and divisions were downloaded from census.gov at:
https://www2.census.gov/geo/pdfs/maps-data/maps/reference/us_regdiv.pdf
}
\seealso{
\link{US_census}
}
\keyword{datasets}
