% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_states_geoms.R
\docType{data}
\name{us_states_geoms}
\alias{us_states_geoms}
\title{Various projections of the states of the USA}
\format{An object of class \code{list} of length 4.}
\source{
Shapefiles from: \url{https://www.census.gov/geo/maps-data/data/cbf/cbf_counties.html}
}
\usage{
us_states_geoms
}
\description{
All geometries were obtained from the US Census website (low resolution,
20m = 1:20,000,000) and subsequently processed as needed.
}
\examples{
library(ggplot2)

# lower 48 states only
ggplot(us_states_geoms$lower48) + geom_sf(fill = "#56B4E9", color = "grey30", size = 0.3, alpha = 0.5)
# true Albers projection, with AK and HI in their correct locations
ggplot(us_states_geoms$true_albers) + geom_sf(fill = "#56B4E9", color = "grey30", size = 0.3, alpha = 0.5)
# standard US Albers map, with AK artificially small
ggplot(us_states_geoms$us_albers) + geom_sf(fill = "#56B4E9", color = "grey30", size = 0.3, alpha = 0.5)
# revised US Albers map, with AK at its original size
ggplot(us_states_geoms$albers_revised) + geom_sf(fill = "#56B4E9", color = "grey30", size = 0.3, alpha = 0.5)
}
\keyword{datasets}
