% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_turbines.R
\docType{data}
\name{wind_turbines}
\alias{wind_turbines}
\title{Locations of wind turbines in the U.S.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 58185 rows and 24 columns.}
\source{
United States Wind Turbine Database (USWTDB)
\url{https://eerscmap.usgs.gov/uswtdb/data/}
}
\usage{
wind_turbines
}
\description{
Data were obtained from the United States Wind Turbine Database.
}
\examples{
library(sf)

m <- cbind(wind_turbines$xlong, wind_turbines$ylat)
mproj <- rgdal::project(m, proj = "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs")
wind_turbines$x <- mproj[,1]
wind_turbines$y <- mproj[,2]

lower48 <- us_states_geoms$lower48
ggplot(lower48) + 
  geom_sf(fill = "#56B4E9", color = "grey30", size = 0.3, alpha = 0.5) +
  geom_point(
    data = filter(wind_turbines, xlong > -130, ylat > 20),
    aes(x = x, y = y),
    color = "black",
    size = 0.1
  )
}
\keyword{datasets}
