% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_log10.R
\name{label_log10}
\alias{label_log10}
\title{Draw labels in log-10 format}
\usage{
label_log10(x)
}
\arguments{
\item{x}{The value to turn into a label}
}
\description{
Draw labels in log-10 format
}
\examples{
library(ggplot2)

ggplot(US_census, aes(pop2000, pop2010)) + 
  geom_point() + 
  scale_x_log10(labels = label_log10) +
  scale_y_log10(labels = label_log10)
}
