% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_figure.R
\name{pad_figure}
\alias{pad_figure}
\title{Pad figure}
\usage{
pad_figure(p, fig.width = 5.5, page.width = 8.5, unit = "in")
}
\description{
Place whitespace to left and right of figure so that the figure spans the entire page width.
}
