% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp}
\alias{stamp}
\alias{stamp_good}
\alias{stamp_bad}
\alias{stamp_ugly}
\title{Stamp plots with a label, such as good, bad, or ugly.}
\usage{
stamp(p, color, alpha, label)

stamp_good(p)

stamp_bad(p)

stamp_ugly(p)
}
\arguments{
\item{p}{The plot to stamp}

\item{color}{The color of the stamp}

\item{alpha}{Transparency level of the stamp}

\item{label}{The text label used for the stamp}
}
\description{
Stamp plots with a label, such as good, bad, or ugly.
}
\examples{
library(ggplot2)

p <- ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width, 
                      color = factor(Petal.Width))) +
  geom_point()
  
stamp_bad(p + guides(color = "none"))
stamp_ugly(p)
}
